requirements.txt
asgiref==3.8.1
click==8.1.7
Django==4.2.17
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
importlib-metadata==6.11.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.2
pillow==11.0.0
pip==24.3.1
pipdeptree==2.24.0
psycopg2-binary==2.9.10
PyJWT==1.7.1  # Warning: This is outdated, recommend >=2.0.0 for compatibility
pytz==2023.4
setuptools==67.8.0
sqlparse==0.4.4
typing_extensions==4.12.2
Werkzeug==2.3.8
wheel==0.45.1
zipp==3.21.0
